Highlight text on desktop when mouse drag is off
When turning off mousedrag, the text in the carousel would still not highlight. I found in the code where the disable events option is and saw that if touch is active and mouse is not it will still disabled highlighting.  Adding in a browser check then checking for the browser touch options as well as the mouse drag should fix this issue.